Add files via upload
This is a program I made for my internship at Rock Engineering. It cycles through a folder of geo-jobsm taking the GPS Coordinates or address of each job and writing them to a text file. This makes it far easier to place pins for each job location in Google Earth.